+++ .fossa.yml
@@ -51,7 +51,7 @@ targets:
       target: ':testing:agent-for-testing'
     - type: gradle
       path: ./
-      target: ':instrumentation:activej-http:javaagent'
+      target: ':instrumentation:activej-http-6.0:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:alibaba-druid-1.0:javaagent'
