+++ settings.gradle.kts
@@ -128,7 +128,7 @@ include(":smoke-tests-otel-starter:spring-boot-reactive-2")
 include(":smoke-tests-otel-starter:spring-boot-reactive-3")
 include(":smoke-tests-otel-starter:spring-boot-reactive-common")
 
-include(":instrumentation:activej-http:javaagent")
+include(":instrumentation:activej-http-6.0:javaagent")
 include(":instrumentation:akka:akka-actor-2.3:javaagent")
 include(":instrumentation:akka:akka-actor-fork-join-2.5:javaagent")
 include(":instrumentation:akka:akka-http-10.0:javaagent")
